backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.13.0
fastapi==0.121.3
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-ws==0.8.2
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
lmstudio==1.5.0
markupsafe==3.0.3
mpmath==1.3.0
msgspec==0.20.0
networkx==3.6
numpy==2.3.5
openai==2.9.0
packaging==25.0
pillow==12.0.0
pydantic==2.12.4
pydantic-core==2.41.5
pymupdf==1.26.6
pypdf==6.4.0
pypdf2==3.0.1
python-dotenv==1.2.1
python-multipart==0.0.21
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wsproto==1.3.2
